+++ .editorconfig
@@ -1,0 +1,26 @@
+# editorconfig.org
+
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.py]
+indent_size = 4
+
+[*.yml]
+indent_size = 4
+
+[*.php]
+indent_size = 4
+
+[*.json]
+indent_size = 4
