+++ .gitignore
@@ -1,0 +1,11 @@
+# IDE & OS
+.idea/
+.DS_Store
+
+# Project
+vendor
+composer.phar
+composer.lock
+phpunit.xml
+Tests/Controller/App/*/
+Tests/Controller/App/sqlite.db.cache
