+++ CraffftContaoDiscourseSSOBundle.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the Craffft Discourse SSO Bundle.
+ *
+ * (c) Daniel Kiesel <https://github.com/iCodr8>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Craffft\ContaoDiscourseSSOBundle;
+
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class CraffftContaoDiscourseSSOBundle extends Bundle
+{
+}
