+++ Resources/contao/config/config.php
@@ -1,4 +1,4 @@
-<?php if (!defined('TL_ROOT')) die('You can not access this file directly!');
+<?php
 
 /**
  * Discourse module for Contao Open Source CMS 2.x
@@ -20,6 +20,7 @@
  * PHP version 5
  * @copyright  Florian Bender 2015
  * @author     Florian Bender <fb+git@quantumedia.de>
+ * @author     Daniel Kiesel <https://github.com/iCodr8>
  * @package    Discourse
  * @license    AGPLv3 (GNU Affero GPL v3.0)
  * @filesource
@@ -31,7 +32,4 @@
  * FRONT END MODULES
  * -------------------------------------------------------------------------
  */
-$GLOBALS['FE_MOD']['application']['discourseSSOProvider'] = 'ModuleSSOProvider';
-
-
- #EOF
+$GLOBALS['FE_MOD']['application']['discourseSSOProvider'] = '\DiscourseSSO\ModuleSSOProvider';
